requirements.txt
python-telegram-bot==13.14
gspread==5.8.0
google-api-python-client==2.100.0
oauth2client==4.1.3
pytesseract==0.3.10
google-auth-oauthlib==1.0.0
pdf2image==1.16.3
openai==0.27.0